feat: mudança do type do id

+++ usuario/src/main/java/br/com/entrega_mais/usuario/repository/UsuarioRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioModel, Integer> {
+public interface UsuarioRepository extends JpaRepository<UsuarioModel, Long> {
 
     public Optional<UsuarioModel> findByEmail(String email);
 
-    public Optional<UsuarioModel> findById(long id);
+    public Optional<UsuarioModel> findById(Long id);
 }
